mnc/requirements.txt
attrs==21.2.0
click==8.0.1
colorama==0.4.4
et-xmlfile==1.1.0
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
numpy==1.20.3
openpyxl==3.0.7
pandas==1.2.4
pymongo==3.11.4
pyreadstat==1.1.2
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
scipy==1.7.0
simplejson==3.17.2
six==1.16.0
spss-converter==0.1.1
validator-collection==1.5.0
Werkzeug==2.0.1
